⭐ FEAT : 검색 시 가게 리스트 출력
Consist of 검색 시 가게 리스트 출력 기능 구현

RESOLVES : #S08P22A307-226

+++ backend/nfast/src/main/java/com/nft/nfast/entity/user/User.java
@@ -17,7 +17,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger userSequence;
+    private long userSequence;
 
     @Column(nullable = false)
     private String userWallet;
@@ -30,7 +30,7 @@ public class User {
     private String userNickname;
 
     @Builder
-    public User(BigInteger userSequence, String userWallet, String userImage, String userNickname) {
+    public User(long userSequence, String userWallet, String userImage, String userNickname) {
         this.userSequence = userSequence;
         this.userWallet = userWallet;
         this.userImage = userImage;
